# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atlas Sanctum Phase 0 API
  version: "0.1.0"
servers:
  - url: http://localhost:4000
paths:
...
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
...